refactor: Participant의 패키지명 변경
- Participant의 패키지명을 ParticipantInfo에서 Participant로 변경
