+++ src/main/java/com/genius/gitget/challenge/certification/service/CertificationService.java
@@ -19,8 +19,8 @@ import com.genius.gitget.challenge.instance.service.InstanceProvider;
 import com.genius.gitget.challenge.item.domain.ItemCategory;
 import com.genius.gitget.challenge.item.domain.UserItem;
 import com.genius.gitget.challenge.item.service.UserItemProvider;
-import com.genius.gitget.challenge.participantinfo.domain.Participant;
-import com.genius.gitget.challenge.participantinfo.service.ParticipantProvider;
+import com.genius.gitget.challenge.participant.domain.Participant;
+import com.genius.gitget.challenge.participant.service.ParticipantProvider;
 import com.genius.gitget.challenge.user.domain.User;
 import com.genius.gitget.global.file.dto.FileResponse;
 import com.genius.gitget.global.file.service.FilesService;
