+++ src/main/java/com/genius/gitget/challenge/participant/domain/JoinResult.java
@@ -1,4 +1,4 @@
-package com.genius.gitget.challenge.participantinfo.domain;
+package com.genius.gitget.challenge.participant.domain;
 
 import lombok.Getter;
