+++ src/main/java/com/genius/gitget/challenge/participant/domain/JoinStatus.java
@@ -1,0 +1,5 @@
+package com.genius.gitget.challenge.participant.domain;
+
+public enum JoinStatus {
+    NO, YES
+}
