--- src/main/java/com/genius/gitget/challenge/participantinfo/domain/JoinStatus.java
@@ -1,5 +1,0 @@
-package com.genius.gitget.challenge.participantinfo.domain;
-
-public enum JoinStatus {
-    NO, YES
-}
